separated edit incident dialog to separate file
