+++ src/components/medicals/incidents/Incident.js
@@ -11,15 +11,7 @@ import StarBorderIcon from "@mui/icons-material/StarBorder";
 import StarIcon from "@mui/icons-material/Star";
 import IconButton from "@mui/material/IconButton";
 import DeleteIcon from "@mui/icons-material/Delete";
-import TextField from "@mui/material/TextField";
-import Dialog from "@mui/material/Dialog";
-import DialogActions from "@mui/material/DialogActions";
-import DialogContent from "@mui/material/DialogContent";
-import DialogTitle from "@mui/material/DialogTitle";
-import FormControl from "@mui/material/FormControl";
-import InputLabel from "@mui/material/InputLabel";
-import Select from "@mui/material/Select";
-import MenuItem from "@mui/material/MenuItem";
+import { EditIncidentDialog } from "./EditIncidentDialog";
 
 export const Incident = ({
     incident,
@@ -29,20 +21,15 @@ export const Incident = ({
     incidentTypes,
     dashboardView
 }) => {
-    const [editedIncident, setEditedIncident] = useState(incident);
-    const [open, setOpen] = useState(false);
 
-    useEffect(() => {
-        setEditedIncident(incident);
-    }, []);
+    const [open, setOpen] = useState(false);
 
     const starUnstar = (event) => {
         event.stopPropagation();
-        const copy = { ...editedIncident };
-        copy.starred = !editedIncident.starred;
+        const copy = { ...incident };
+        copy.starred = !incident.starred;
         delete copy.incidentType;
         delete copy.pet;
-        setEditedIncident(copy);
         MedicalRepository.editIncident(copy).then(() => syncIncidents());
     };
 
@@ -53,20 +40,8 @@ export const Incident = ({
         );
     };
 
-    const editPetIncident = (event) => {
-        event.preventDefault();
-        handleClose();
-        const copy = { ...editedIncident };
-        delete copy.incidentType;
-        MedicalRepository.editIncident(copy).then(() => syncIncidents());
-    };
-
     const handleClickOpen = () => {
         setOpen(true);
-    };
-
-    const handleClose = () => {
-        setOpen(false);
     };
 
     return (
@@ -120,83 +95,7 @@ export const Incident = ({
                     <Button onClick={handleClickOpen}>Edit</Button>
                 </AccordionDetails>
             </Accordion>
-            <Dialog open={open} onClose={handleClose}>
-                <form onSubmit={editPetIncident}>
-                    <DialogTitle>Edit Incident Record</DialogTitle>
-                    <DialogContent>
-                        <TextField
-                            margin="dense"
-                            id="name"
-                            label="name"
-                            value={editedIncident.name}
-                            required
-                            type="text"
-                            onChange={(event) => {
-                                const copy = { ...editedIncident };
-                                copy.name = event.target.value;
-                                setEditedIncident(copy);
-                            }}
-                        />
-                        <TextField
-                            margin="dense"
-                            id="description"
-                            label="description"
-                            value={editedIncident.description}
-                            required
-                            type="text"
-                            onChange={(event) => {
-                                const copy = { ...editedIncident };
-                                copy.description = event.target.value;
-                                setEditedIncident(copy);
-                            }}
-                        />
-                        <TextField
-                            margin="dense"
-                            id="date"
-                            label="date"
-                            value={editedIncident.date}
-                            required
-                            type="date"
-                            onChange={(event) => {
-                                const copy = { ...editedIncident };
-                                copy.date = event.target.value;
-                                setEditedIncident(copy);
-                            }}
-                        />
-                        <FormControl required sx={{ m: 1, minWidth: 225 }}>
-                            <InputLabel id="shot-label">
-                                Incident type
-                            </InputLabel>
-                            <Select
-                                labelId="incident-type-label"
-                                id="incident-type"
-                                value={editedIncident.incidentTypeId}
-                                label="incident-type"
-                                onChange={(event) => {
-                                    const copy = { ...editedIncident };
-                                    copy.incidentTypeId = parseInt(
-                                        event.target.value
-                                    );
-                                    setEditedIncident(copy);
-                                }}
-                            >
-                                {incidentTypes.map((type) => (
-                                    <MenuItem
-                                        key={`incident-type--${type.id}`}
-                                        value={type.id}
-                                    >
-                                        {type.label}
-                                    </MenuItem>
-                                ))}
-                            </Select>
-                        </FormControl>
-                    </DialogContent>
-                    <DialogActions>
-                        <Button onClick={handleClose}>Cancel</Button>
-                        <Button type="submit">Save</Button>
-                    </DialogActions>
-                </form>
-            </Dialog>
+            <EditIncidentDialog open={open} setOpen={setOpen} incident={incident} syncIncidents={syncIncidents} incidentTypes={incidentTypes}/>
         </>
     );
 };
